fix(migrations): make 037 safe on fresh databases
037 assumed team_users.updated_at/password_reset_expires/invite_expires and the
sessions table already existed — but nothing in migrations/ creates them:
lib/auth.ts adds them lazily at RUNTIME, and migrations run at boot before any
runtime code. So a from-scratch database (new environment, disaster recovery,
local dev) failed 037 with 'column "updated_at" does not exist' and the
container crash-looped. Production never hit this because its columns predate
the file, and since schema_migrations already records 037 as applied there,
this rewrite never re-runs in production — the change is fresh-DB-only.

  * 037 now creates the auth columns and sessions table (as BIGINT) if absent,
    then widens to BIGINT unconditionally; every statement is idempotent
  * lib/auth.ts lazy-create switched INTEGER→BIGINT for the same columns — it
    was still creating the exact 32-bit shape whose Date.now() overflow 037
    exists to fix, and would have reintroduced it anywhere it ran first

Verified against throwaway Postgres 16: (1) all 47 migrations apply cleanly on
a fresh DB in boot order — previously stopped at 037; (2) 037 applied to a
simulated pre-037 schema with INTEGER columns widens them without error;
(3) re-running 037 is a no-op; (4) resulting types are all bigint and
millisecond-epoch inserts succeed.

+++ lib/auth.ts
@@ -26,12 +26,18 @@ export async function ensureAuthSchema(): Promise<void> {
 
   // Create the sessions table if it doesn't exist. Uses IF NOT EXISTS so it's
   // idempotent across container restarts and dev hot-reloads.
+  //
+  // Timestamp columns are BIGINT, and must stay BIGINT: they store Date.now()
+  // (~1.78T in 2026), which overflows INTEGER's 2.1B max. This lazy-create used
+  // to say INTEGER, which is the exact bug migration 037 was written to fix —
+  // any environment where this ran before 037's rewrite got overflowing
+  // sessions. Keep this definition in lockstep with migrations/037.
   await db.query(`
     CREATE TABLE IF NOT EXISTS sessions (
       token      TEXT PRIMARY KEY,
       user_id    TEXT NOT NULL,
-      created_at INTEGER NOT NULL,
-      expires_at INTEGER NOT NULL,
+      created_at BIGINT NOT NULL,
+      expires_at BIGINT NOT NULL,
       FOREIGN KEY (user_id) REFERENCES team_users(id) ON DELETE CASCADE
     )
   `);
@@ -39,13 +45,14 @@ export async function ensureAuthSchema(): Promise<void> {
   // Add auth columns to team_users — silently skip if they already exist.
   // Postgres doesn't support ADD COLUMN IF NOT EXISTS in older versions, so we
   // use a DO block to catch the duplicate-column error rather than failing the whole call.
+  // BIGINT for the same Date.now() reason as sessions above (was INTEGER; see 037).
   const authColumns: [string, string][] = [
     ['password_hash',           'TEXT'],
-    ['updated_at',              'INTEGER'],
+    ['updated_at',              'BIGINT'],
     ['password_reset_token',    'TEXT'],
-    ['password_reset_expires',  'INTEGER'],
+    ['password_reset_expires',  'BIGINT'],
     ['invite_token',            'TEXT'],
-    ['invite_expires',          'INTEGER'],
+    ['invite_expires',          'BIGINT'],
   ];
 
   for (const [col, type] of authColumns) {
